Fixed several bugs in set_register_value
There were several instances of not shifting correctly, which totally messed things up.

Go ahead and set the value immediately if the register is the full width register.

Normalize the value to an unsigned value the size of the register to be set, which allows users to provide negative numbers